Fix permission checks to work with more then one user
Previously, the permissions would check if any condition failed, rather
then checking if any condition succeeded. This made it practically
impossible to setup fine-grained permissions per-user or per-command.

Now, the check will succeed if any permission entry matches, either for
a role or a user id, as well as the command.